requirements.txt
kafka-python==2.0.2
pymongo==4.6.1
pytrends==4.9.2
beautifulsoup4==4.12.2
requests==2.32.3